third_party/WebKit/LayoutTests/bluetooth/getCharacteristics/characteristics-not-found-with-uuid.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../resources/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  testRunner.setBluetoothMockDataSet('HeartRateAdapter');
  let expected = new DOMException(
    'No Characteristics with specified UUID found in Service.',
    'NotFoundError');
  return requestDeviceWithKeyDown({filters: [{services: ['heart_rate']}]})
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService('heart_rate'))
    .then(service => assert_promise_rejects_with_message(
      service.getCharacteristics('battery_level'), expected));
}, 'Request for absent characteristics with UUID. Reject with NotFoundError.');
</script>
